+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add module ``interfaces``
   - add ``data_exchange`` interfaces, to pass calender entries or list of them
     between different components of the application
+- add a generic plugin api
+- add a plugin api to provide events